+++ sdk/nodejs/apps/v1/Deployment.ts
@@ -25,6 +25,10 @@ import { getVersion } from "../../version";
      *    'NewReplicaSetAvailable'. For generation <= 1, this status field does not exist,
      *    because it doesn't do a rollout (i.e., it simply creates the Deployment and
      *    corresponding ReplicaSet), and therefore there is no rollout to mark as 'Progressing'.
+     * 
+     * If the Deployment has not reached a Ready state after 5 minutes, it will
+     * time out and mark the resource update as Failed. You can override the default timeout value
+     * by setting the 'pulumi.com/timeoutSeconds' as a '.metadata.annotation' on the resource.
      */
     export class Deployment extends pulumi.CustomResource {
       /**