+++ sdk/nodejs/apps/v1beta1/StatefulSet.ts
@@ -22,6 +22,10 @@ import { getVersion } from "../../version";
      * 1. The value of 'spec.replicas' matches '.status.replicas', '.status.currentReplicas',
      *    and '.status.readyReplicas'.
      * 2. The value of '.status.updateRevision' matches '.status.currentRevision'.
+     * 
+     * If the StatefulSet has not reached a Ready state after 5 minutes, it will
+     * time out and mark the resource update as Failed. You can override the default timeout value
+     * by setting the 'pulumi.com/timeoutSeconds' as a '.metadata.annotation' on the resource.
      */
     export class StatefulSet extends pulumi.CustomResource {
       /**