+++ sdk/nodejs/networking/v1beta1/Ingress.ts
@@ -19,6 +19,10 @@ import { getVersion } from "../../version";
      * 2.  Endpoint objects exist with matching names for each Ingress path (except when Service
      *     type is ExternalName).
      * 3.  Ingress entry exists for '.status.loadBalancer.ingress'.
+     * 
+     * If the Ingress has not reached a Ready state after 10 minutes, it will
+     * time out and mark the resource update as Failed. You can override the default timeout value
+     * by setting the 'pulumi.com/timeoutSeconds' as a '.metadata.annotation' on the resource.
      */
     export class Ingress extends pulumi.CustomResource {
       /**