+++ sdk/python/pulumi_kubernetes/apps/v1/StatefulSet.py
@@ -26,6 +26,10 @@ class StatefulSet(pulumi.CustomResource):
     1. The value of 'spec.replicas' matches '.status.replicas', '.status.currentReplicas',
        and '.status.readyReplicas'.
     2. The value of '.status.updateRevision' matches '.status.currentRevision'.
+    
+    If the StatefulSet has not reached a Ready state after 5 minutes, it will
+    time out and mark the resource update as Failed. You can override the default timeout value
+    by setting the 'pulumi.com/timeoutSeconds' as a '.metadata.annotation' on the resource.
     """
 
     def __init__(self, resource_name, opts=None, metadata=None, spec=None, status=None, __name__=None, __opts__=None):