+++ sdk/python/pulumi_kubernetes/apps/v1beta1/Deployment.py
@@ -33,6 +33,10 @@ class Deployment(pulumi.CustomResource):
        'NewReplicaSetAvailable'. For generation <= 1, this status field does not exist,
        because it doesn't do a rollout (i.e., it simply creates the Deployment and
        corresponding ReplicaSet), and therefore there is no rollout to mark as 'Progressing'.
+    
+    If the Deployment has not reached a Ready state after 5 minutes, it will
+    time out and mark the resource update as Failed. You can override the default timeout value
+    by setting the 'pulumi.com/timeoutSeconds' as a '.metadata.annotation' on the resource.
     """
 
     def __init__(self, resource_name, opts=None, metadata=None, spec=None, status=None, __name__=None, __opts__=None):